requirements-dev.txt
django-modeltranslation==0.18.12
django-extensions==3.2.3
django-jet==1.0.8

# Testing and development tools
djangorestframework==3.15.2
pytest==8.3.4
pytest-django==4.9.0
